removed unused lines and common transfomer functions
