+++ configs/classification/imagenet/edgevit/imagenet_edgeVIT_xs_jpg.py
@@ -1,5 +1,4 @@
 _base_ = './EdgeVit_b512x8_300e_jpg.py'
-# model settings
 model = dict(
     type='Classification',
     backbone=dict(
@@ -17,11 +16,16 @@ model = dict(
         with_avg_pool=True,
         in_channels=384,
         loss_config=dict(
-            type='CrossEntropyLossWithLabelSmooth', label_smooth=0),
+            type='CrossEntropyLossWithLabelSmooth', label_smooth=0.1),
     ))
 
+# input data settings
 data = dict(
-    imgs_per_gpu=256,  # total 256
+    imgs_per_gpu=256,
     workers_per_gpu=10,
     use_repeated_augment_sampler=True,
 )
+
+# optimizer
+update_interval = 4
+optimizer_config = dict(update_interval=update_interval)
